requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0

# YOLO
ultralytics>=8.0.0

# Segmentation
albumentations>=1.3.0
segmentation-models-pytorch>=0.3.0

# LLM
transformers>=4.35.0
accelerate>=0.24.0

# Utilities
scikit-learn>=1.3.0
scipy>=1.11.0
